Separate PPU register tests into a dedicated suite
There are quite a few nuances associated with reading and writing to these registers that warrant a dedicated suite of tests to improve readability. This change does not modify any of the existing tests.
